Fix some int props that should be ushort; allow for blank values in config to overwrite existing
git-svn-id: svn://localhost/Mailman@739 bc798c11-6c20-9a46-872d-b4c02895e90f

+++ MailmanSharp/Sections/AutoResponderSection.cs
@@ -17,7 +17,7 @@ namespace MailmanSharp.Sections
         public List<string> AutoresponseAdminText { get; set; }
         public AutorespondRequestsOption AutorespondRequests { get; set; }
         public List<string> AutoresponseRequestText { get; set; }
-        public int AutoresponseGraceperiod { get; set; }
+        public ushort AutoresponseGraceperiod { get; set; }
 
 
         public AutoResponderSection(MailmanList list) : base(list) { }
